pyproject.toml
[tool.poetry]
name = "trood_test-master"
version = "0.1.0"
description = ""
authors = ["Spartan327 <vbutovchenko@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
flake8 = "^3.9.2"
psycopg2 = "^2.9.1"
pytest-django = "^4.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
